+++ src/containers/Sidebar/NavbarBrand.js
@@ -1,14 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
 import { Link } from 'react-router-dom';
 
 import { getBrandImage } from '../../components/branding/BrandImage';
+import { useAppSelector } from 'asab_webui_components';
 
 const NavbarBrand = ( props ) => {
 	const [ brandImage, setBrandImage ] = useState({});
-	const theme = useSelector(state => state.theme);
-	const title = useSelector(state => state.config?.title);
-	const isSidebarCollapsed = useSelector(state => state.sidebar?.isSidebarCollapsed);
+	const theme = useAppSelector(state => state.theme);
+	const title = useAppSelector(state => state.config?.title);
+	const isSidebarCollapsed = useAppSelector(state => state.sidebar?.isSidebarCollapsed);
 
 	useEffect(() => {
 		setBrandImage(getBrandImage(props, theme));