+++ src/modules/auth/index.js
@@ -146,7 +146,7 @@ export default class AuthModule extends Module {
 				}
 
 				// Validate resources of items and children in navigation (resource validation depends on tenant)
-				if (this.App.Store && this.App.Services.TenantService && (this.UserInfo !== null)) {
+				if (this.App.AppStore && this.App.Services.TenantService && (this.UserInfo !== null)) {
 					let currentTenant = this.App.Services.TenantService.getCurrentTenant();
 					let resources = this.UserInfo.resources ? this.UserInfo.resources[currentTenant] : [];
 					/*
@@ -161,10 +161,8 @@ export default class AuthModule extends Module {
 						return;
 					}
 
-					if (this.App.Store != null) {
-						this.App.Store.dispatch({ type: types.AUTH_RESOURCES, resources: resources });
-					}
-					await this.validateNavigation();
+					this.App?.AppStore?.dispatch?.({ type: types.AUTH_RESOURCES, resources: resources });
+					this.validateNavigation({resources});
 				}
 
 				if (this.UserInfo != null) {
@@ -228,9 +226,9 @@ export default class AuthModule extends Module {
 			mockParams["tenants"] = Object.values(mockParams.tenants)
 		}
 
-		if (this.App.Store != null) {
-			this.App.Store.dispatch({ type: types.AUTH_USERINFO, payload: mockParams });
-			this.App.Store.dispatch({ type: types.AUTH_RESOURCES, resources: mockParams["resources"] });
+		if (this.App.AppStore) {
+			this.App.AppStore.dispatch?.({ type: types.AUTH_USERINFO, payload: mockParams });
+			this.App.AppStore.dispatch?.({ type: types.AUTH_RESOURCES, resources: mockParams["resources"] });
 		}
 
 		/** Check for TenantService and pass tenants list obtained from userinfo */
@@ -258,20 +256,18 @@ export default class AuthModule extends Module {
 		});
 	}
 
-	// TODO: reconsider removing async/await with promise
-	async validateNavigation() {
-		const state = this.App.Store.getState();
+	validateNavigation({resources}) {
+		const state = this.App.AppStore.getState();
 		let navItems = state.navigation?.navItems;
-		let resources = state.auth?.resources;
 		let authorizedNavItems = [];
-		navItems && await Promise.all(navItems.map(async (itm, idx) => {
+		navItems && navItems.map((itm, idx) => {
 			if (resources.indexOf('authz:superuser') !== -1) {
 				// If user is superuser, validate every navigation item
 				authorizedNavItems.push(itm);
 			} else {
 				if (itm.resource) {
 					if (itm.children) {
-						itm.children = await this._validateChildrenNav(itm, resources);
+						itm.children = this._validateChildrenNav(itm, resources);
 					}
 					// Item validation
 					let access_auth = this._validateItemNav(itm.resource, resources);
@@ -280,29 +276,29 @@ export default class AuthModule extends Module {
 					}
 				} else {
 					if (itm.children) {
-						itm.children = await this._validateChildrenNav(itm, resources);
+						itm.children = this._validateChildrenNav(itm, resources);
 						if (itm.children.length > 0) {
 							authorizedNavItems.push(itm);
 						}
 					}
 				}
 			}
-		}))
-		this.App.Store.dispatch({ type: SET_NAVIGATION_ITEMS, navItems: authorizedNavItems });
+		})
+		this.App.AppStore.dispatch?.({ type: SET_NAVIGATION_ITEMS, navItems: authorizedNavItems });
 	}
 
 	// Validate sidebar's item children
-	async _validateChildrenNav(itm, resources) {
+	_validateChildrenNav(itm, resources) {
 		// Item's children validation
 		let authorizedNavChildren = [];
-		await Promise.all(itm.children.map(async (child, id) => {
+		itm.children.map((child, id) => {
 			if (child.resource) {
 				let access_auth = this._validateItemNav(child.resource, resources);
 				if (access_auth == true) {
 					authorizedNavChildren.push(child);
 				}
 			}
-		}))
+		})
 		return authorizedNavChildren;
 	}
 
@@ -341,16 +337,16 @@ export default class AuthModule extends Module {
 		catch (err) {
 			console.error("Failed to update user info", err);
 			this.UserInfo = null;
-			if (this.App.Store != null) {
-				this.App.Store.dispatch({ type: types.AUTH_USERINFO, payload: this.UserInfo });
+			if (this.App.AppStore) {
+				this.App.AppStore.dispatch?.({ type: types.AUTH_USERINFO, payload: this.UserInfo });
 			}
 			return false;
 		}
 
 		this.UserInfo = response.data;
 		this.SessionExpiration = response.data?.exp;
-		if (this.App.Store != null) {
-			this.App.Store.dispatch({ type: types.AUTH_USERINFO, payload: this.UserInfo });
+		if (this.App.AppStore) {
+			this.App.AppStore.dispatch?.({ type: types.AUTH_USERINFO, payload: this.UserInfo });
 		}
 
 		/** Check for TenantService and pass tenants list obtained from userinfo */
@@ -486,10 +482,10 @@ export default class AuthModule extends Module {
 					// Stop further checks
 					clearTimeout(this.sessionValidationInterval);
 					this.sessionValidationInterval = null;
-					this.App.addAlert("info", "ASABAuthModule|Your session has expired.", 3600 * 1000, true, (alert, store) => <SessionExpirationAlert alert={alert} store={store} />);
+					this.App.addAlert("info", "ASABAuthModule|Your session has expired.", 3600 * 1000, true, (alert) => <SessionExpirationAlert alert={alert} />);
 					// Disable UI elements
-					if (this.App.Store) {
-						this.App.Store.dispatch({ type: types.AUTH_SESSION_EXPIRATION, sessionExpired: true });
+					if (this.App.AppStore) {
+						this.App.AppStore.dispatch?.({ type: types.AUTH_SESSION_EXPIRATION, sessionExpired: true });
 
 						// Disable UI elements
 						[...document.querySelectorAll('#app-sidebar .nav-link, [class^="btn"]:not(.alert-button), [class*=" btn"]:not(.alert-button), .btn-group a, .page-item, input, select')].forEach(i => {