+++ CHANGELOG.md
@@ -1,8 +1,20 @@
 # CHANGELOG for ASAB WebUI Shell
 
-## 25.3.1
+## 26.2.1
 
 - Make AttentionRequired service a reusable module. Then it is initialized only in desired application where it is imported. (#29)
+
+## 26.1.3
+
+- Refactor Application store with createStore() from asab_webui_components to make global variables dynamic and fix issues with not updating the AppStore state dynamically. (#31)
+
+## 26.1.2
+
+- Import FlowbiteIllustration from `asab_webui_components` into informational screens (#27)
+
+## 26.1.1
+
+- Upgrade to react v19, remove redux dependency and replace it with custom redux-like context and AppStore, remove unused flag header (#17)
 
 ## 25.2.8